+++ arctic_office_projects_api/models.py
@@ -21,11 +21,14 @@ faker.add_provider(ProjectProvider)
 faker.add_provider(PersonProvider)
 faker.add_provider(ProfileProvider)
 
-static_project_duration = DateRange(date(2013, 3, 1), date(2016, 10, 1))
+static_grant_duration = DateRange(date(2013, 3, 1), date(2016, 10, 1))
+static_project_duration = static_grant_duration
 
 static_project_nid = '01D5M0CFQV4M7JASW7F87SRDYB'
 static_participant_nid = '01D5T4N25RV2062NVVQKZ9NBYX'
 static_person_nid = '01D5MHQN3ZPH47YVSVQEVB0DAE'
+static_grant_nid = '01D6T4HYAVK5SJFD7NWJRMBZ4Z'
+static_allocation_nid = '01D6T4QQNDBJTSEVXESNXD3AN0'
 
 
 class Project(db.Model):
@@ -44,6 +47,7 @@ class Project(db.Model):
     project_duration = db.Column(postgresql.DATERANGE(), nullable=False)
 
     participants = db.relationship("Participant", back_populates="project")
+    allocations = db.relationship("Allocation", back_populates="project")
 
     def __repr__(self):
         return f"<Project { self.neutral_id }>"
@@ -798,3 +802,185 @@ class Participant(db.Model):
                         )
 
                         db.session.add(co_investigator)
+
+
+class GrantStatus(Enum):
+    """
+    Represents the various states of a research grant
+    """
+    Accepted = {
+        'title': 'accepted'
+    }
+    Active = {
+        'title': 'active'
+    }
+    Approved = {
+        'title': 'approved'
+    }
+    Authorised = {
+        'title': 'active'
+    }
+    Closed = {
+        'title': 'closed'
+    }
+
+
+class GrantCurrency(Enum):
+    """
+    Represents the various currencies of a research grant
+    """
+    GBP = {
+        'iso_4217_code': 'GBP',
+        'major_symbol': '£'
+    }
+    EUR = {
+        'iso_4217_code': 'EUR',
+        'major_symbol': '€'
+    }
+    USD = {
+        'iso_4217_code': 'USD',
+        'major_symbol': '$'
+    }
+
+
+class Grant(db.Model):
+    """
+    Represents information about a research grant
+    """
+    __tablename__ = 'grants'
+    id = db.Column(db.Integer, primary_key=True)
+    neutral_id = db.Column(db.String(32), unique=True, nullable=False, index=True)
+    reference = db.Column(db.Text(), unique=True, nullable=False)
+    title = db.Column(db.Text(), nullable=False)
+    abstract = db.Column(db.Text(), nullable=True)
+    website = db.Column(db.Text(), nullable=True)
+    publications = db.Column(postgresql.ARRAY(db.Text(), dimensions=1, zero_indexes=True), nullable=True)
+    duration = db.Column(postgresql.DATERANGE(), nullable=False)
+    status = db.Column(db.Enum(GrantStatus), nullable=True)
+    total_funds = db.Numeric(24, 2)
+    total_funds_currency = db.Column(db.Enum(GrantCurrency), nullable=True)
+    in_direct_funds = db.Numeric(24, 2)
+    in_direct_funds_currency = db.Column(db.Enum(GrantCurrency), nullable=True)
+
+    allocations = db.relationship("Allocation", back_populates="grant")
+
+    def __repr__(self):
+        return f"<Grant { self.neutral_id } ({ self.reference })>"
+
+    @staticmethod
+    def seed(*, quantity: int = 1):
+        """
+        Populate database with mock/fake data
+
+        By default, a single, static, resource will be added to allow testing against a predictable/stable instance.
+        Additional instances are created randomly using Faker.
+
+        The quantity parameter is treated as a target number of resources to add, as Faker is unaware of unique
+        constraints, and may use the same values twice. Resources with duplicate values are discarded resulting in
+        fewer resources being added. For example, if 250 resources are requested, only 246 may be unique.
+
+        :type quantity: int
+        :param quantity: target number of Grant resources to create
+        """
+        if not db.session.query(exists().where(Grant.neutral_id == static_grant_nid)).scalar():
+            static_grant = Grant(
+                neutral_id=static_grant_nid,
+                reference='NE/I028769/1',
+                title='Aerosol-Cloud Coupling And Climate Interactions in the Arctic',
+                abstract="The Arctic climate is changing twice as fast as the global average and these dramatic "
+                         "changes are evident in the decreases in sea ice extent over the last few decades. The "
+                         "lowest sea ice cover to date was recorded in 2007 and recent data suggests sea ice cover "
+                         "this year may be even lower. Clouds play a major role in the Arctic climate and therefore "
+                         "influence the extent of sea ice, but our understanding of these clouds is very poor. Low "
+                         "level, visually thick, clouds in much of the world tend to have a cooling effect, because "
+                         "they reflect sunlight back into space that would otherwise be absorbed at the surface. "
+                         "However, in the Arctic this albedo effect is not as important because the surface, often "
+                         "being covered in snow and ice, is already highly reflective and Arctic clouds therefore "
+                         "tend to warm instead of cooling. Warming in the Arctic can, in turn, lead to sea ice "
+                         "break-up which exposes dark underlying sea water. The sea water absorbs more of the sun's "
+                         "energy, thus amplifying the original warming. Hence, small changes in cloud properties or "
+                         "coverage can lead to dramatic changes in the Arctic climate; this is where the proposed "
+                         "research project comes in. \n A large portion of clouds, including those found in the Arctic "
+                         "region, are categorized as mixed phase clouds. This means they contain both supercooled "
+                         "water droplets and ice crystals (for a demonstration of supercooled water see: "
+                         "http://www.youtube.com/watch?v=0JtBZGXd5zo). Liquid cloud droplets can exist in a "
+                         "supercooled state well below zero degrees centigrade without freezing. Freezing will, "
+                         "however, be observed if the droplets contain a particle known as an ice nucleus that can "
+                         "catalyze ice formation and growth. Ice formation dramatically alters a cloud's properties "
+                         "and therefore its influence on climate. At lower latitudes, ice nuclei are typically made up "
+                         "of desert dusts, soot or even bacteria. But the composition and source of ice nuclei in the "
+                         "Arctic environment remains a mystery. \n A likely source of ice nuclei in the Arctic is the "
+                         "ocean. Particles emitted at the sea surface, through the action of waves breaking and bubble "
+                         "bursting, may serve as ice nuclei when they are lofted into the atmosphere and are "
+                         "incorporated in cloud droplets. This source of ice nuclei has not yet been quantified. We "
+                         "will be the first to make measurements of ice nuclei in the central Arctic region. We will "
+                         "make measurements of ice nuclei in the surface layers of the sea from a research ship as "
+                         "well as measuring airborne ice nuclei from the BAe-146 research aircraft. \n The sea's "
+                         "surface contains a wide range of bacteria, viruses, plankton and other materials which are "
+                         "ejected into the atmosphere and may cause ice to form. We will use state-of-the-art "
+                         "equipment developed at Leeds to measure how well sea-derived particles and particles sampled "
+                         "in the atmosphere nucleate ice. We will piggy back on a NERC funded project called ACACCIA, "
+                         "which not only represents excellent value for money (since the ship and aircraft are already "
+                         "paid for under ACCACIA), but is a unique opportunity to access this remote region. \n "
+                         "Results from the proposed study will build upon previous work performed in the Murray "
+                         "laboratory and generate quantitative results that can be directly used to improve "
+                         "computer-based cloud, aerosol and climate models. Our results will further our "
+                         "understanding of these mysterious and important mixed phase clouds and, in turn, the global "
+                         "climate.",
+                website='https://gtr.ukri.org/projects?ref=NE%2FI028769%2F1',
+                publications=[
+                    'https://doi.org/10.5194/acp-2018-283',
+                    'https://doi.org/10.5194/acp-15-3719-2015',
+                    'https://doi.org/10.5194/acp-15-5599-2015',
+                    'https://doi.org/10.5194/acp-16-4063-2016'
+                ],
+                duration=static_grant_duration,
+                status=GrantStatus.Closed,
+                total_funds=324282
+            )
+            db.session.add(static_grant)
+
+        if quantity > 1:
+            pass
+
+
+class Allocation(db.Model):
+    """
+    Represents the relationship between an research grant and a research project (i.e. the funding for a project)
+    """
+    __tablename__ = 'allocations'
+    id = db.Column(db.Integer, primary_key=True)
+    neutral_id = db.Column(db.String(32), unique=True, nullable=False, index=True)
+    project_id = db.Column(db.Integer, db.ForeignKey('projects.id'), nullable=False)
+    grant_id = db.Column(db.Integer, db.ForeignKey('grants.id'), nullable=False)
+
+    project = db.relationship("Project", back_populates="allocations")
+    grant = db.relationship("Grant", back_populates="allocations")
+
+    def __repr__(self):
+        return f"<Allocation { self.neutral_id } ({ self.grant.neutral_id }:{ self.project.neutral_id })>"
+
+    @staticmethod
+    def seed(*, quantity: int = 1):
+        """
+        Populate database with mock/fake data
+
+        By default, a single, static, resource will be added to allow testing against a predictable/stable instance.
+        Additional instances are created randomly using Faker.
+
+        The quantity parameter is treated as a trigger to randomly assign a grant to each project - the actual number
+        for this parameter does not matter, providing it is greater than 1.
+
+        :type quantity: int
+        :param quantity: target number of Allocation resources to create
+        """
+        if not db.session.query(exists().where(Allocation.neutral_id == static_allocation_nid)).scalar():
+            static_allocation = Allocation(
+                neutral_id=static_allocation_nid,
+                project=Project.query.filter_by(neutral_id=static_project_nid).one(),
+                grant=Grant.query.filter_by(neutral_id=static_grant_nid).one()
+            )
+            db.session.add(static_allocation)
+
+        if quantity > 1:
+            pass